refactor(frontend): consolidate tests into src/test/ directory
Move all test files from alongside source files into a mirrored
structure under src/test/:
  src/test/utils/
  src/test/connections/internal/
  src/test/components/

Update all relative imports accordingly.
